Fixed a number of bugs/enhancements found by users.
Thanks to gpminsuk for finding an issue when a propery in a collection
had a 0 value, a null was returned!
Thanks to pascalberge, moved the jinqJs definition file to
scripts\typings\jinqJs.
Thanks to pemn, for finding if a collection has been added using a
from() and then later .delete().at() is called with no parameters, the
internal variable collections is not being cleared.
Thanks to ninety7 for suggesting using string.localeCompare() for
ordering.